Update README.md
Disasters occur often and the most vulnerable are the poor who are exposed to various disaster risks. Majority live in developing countries where the level of disaster preparedness, resources and knowhow is still low. India, like other developing countries of the world, is vulnerable to disaster risks resulting in deaths and loss of property worth millions. Geo-Spatial analysis (GSA) have generated great interest in the field of disaster management and have stimulated rich discussions in the domain. This GSA-enabled mobile application to be used by field workers in the aftermath of disaster. The app will help decision-makers, middle managers/supervisors, and field workers achieve faster restoration and deliver a more efficient post-disaster emergency response. The findings of this project could serve as a guide for emergency restoration teams in similar disaster management scenarios in the future. A B2C AI model, which can tackle large volumes of information during calamities, assist the relief committee and prioritize the information accurately, reducing the response time and damages. Enhanced multi-level precedence and pre-emption based priority relief routing should be implemented in India along with the option to pre-empt ongoing services, if needed in order to ensure efficacy.

The objectives of the project are  Safeguard and make available virtual materials, supplies and equipment to ensure safety.  Reduce the risk of disasters caused by human errors, deliberate destruction and building or equipment failure.  Be prepared to recover from a major Natural catastrophe.  Ensure the organizations ability to continue operating after a disaster.  To provide Minimal human intervention to disseminate genuine news to people and emergency aid  To help various safety management departments in restoration of electricity, supply of food, medical aid etc.  To help Government in analysis of damage after a disaster.